// src/word-checker/dto/requirements.dto.ts
export class UserStylingDto {
  docMode: string;

  docSize: string;

  top?: number;

  bottom?: number;

  left?: number;

  right?: number;

  fromEdgeHeader?: number;

  fromEdgeFooter?: number;

  textStyleFont?: string;
  textStyleSize?: number;
  // mainSectionHeadingFont?: string;
  // mainSectionHeadingSize?: number;
  // mainSectionHeadingStyle?: string;
  // secondLevelHeadingFont?: string;
  // secondLevelHeadingSize?: number;
  // secondLevelHeadingStyle?: string;
  // thirdLevelHeadingFont?: string;
  // thirdLevelHeadingSize?: number;
  // thirdLevelHeadingStyle?: string;
  // fifthLevelHeadingFont?: string;
  // fifthLevelHeadingSize?: number;
}